requirements.txt
selenium
webdriver-manager
google-generativeai